+++ library/src/main/java/eu/kanade/tachiyomi/source/Source.kt
@@ -1,18 +1,21 @@
 package eu.kanade.tachiyomi.source
 
+import eu.kanade.tachiyomi.source.model.FilterList
+import eu.kanade.tachiyomi.source.model.MangasPage
 import eu.kanade.tachiyomi.source.model.Page
 import eu.kanade.tachiyomi.source.model.SChapter
 import eu.kanade.tachiyomi.source.model.SManga
+import eu.kanade.tachiyomi.source.model.SMangaUpdate
 import rx.Observable
 
 /**
  * A basic interface for creating a source. It could be an online source, a local source, etc...
  */
-@Suppress("unused")
+@Suppress("UNUSED")
 interface Source {
 
     /**
-     * Id for the source. Must be unique.
+     * ID for the source. Must be unique.
      */
     val id: Long
 
@@ -22,41 +25,76 @@ interface Source {
     val name: String
 
     /**
-     * Get the updated details for a manga.
-     *
-     * @since extensions-lib 1.4
-     * @param manga the manga to update.
-     * @return the updated manga.
+     * Whether the source has support for latest updates.
      */
-    suspend fun getMangaDetails(manga: SManga): SManga = throw Exception("Stub!")
+    val supportsLatest: Boolean
 
     /**
-     * Get all the available chapters for a manga.
-     *
-     * @since extensions-lib 1.4
-     * @param manga the manga to update.
-     * @return the chapters for the manga.
+     * Returns the list of filters for the source.
      */
-    suspend fun getChapterList(manga: SManga): List<SChapter> = throw Exception("Stub!")
+    fun getFilterList(): FilterList = throw Exception("Stub!")
+
+    /**
+     * Get a page with a list of manga.
+     *
+     * @since tachiyomix 1.6
+     * @param page the page number to retrieve.
+     */
+    suspend fun getPopularManga(page: Int): MangasPage
+
+    /**
+     * Get a page with a list of latest manga updates.
+     *
+     * @since tachiyomix 1.6
+     * @param page the page number to retrieve.
+     */
+    suspend fun getLatestUpdates(page: Int): MangasPage
+
+    /**
+     * Get a page with a list of manga.
+     *
+     * @since tachiyomix 1.6
+     * @param page the page number to retrieve.
+     * @param query the search query.
+     * @param filters the list of filters to apply.
+     */
+    suspend fun getSearchManga(page: Int, query: String, filters: FilterList): MangasPage
+
+    /**
+     * Fetches updated information for a manga.
+     *
+     * Depending on the provided flags or source availability, this may include
+     * updated manga metadata, available chapters, or both.
+     *
+     * @since tachiyomix 1.6
+     * @param manga The manga to fetch updates for.
+     * @param chapters Existing chapters of the manga
+     * @param includeManga Whether to include updated manga details.
+     * @param includeChapters Whether to include available chapters.
+     */
+    suspend fun getMangaUpdate(
+        manga: SManga,
+        chapters: List<SChapter>,
+        includeManga: Boolean,
+        includeChapters: Boolean,
+    ): SMangaUpdate
 
     /**
      * Get the list of pages a chapter has. Pages should be returned
      * in the expected order; the index is ignored.
      *
+     * @since tachiyomix 1.6
      * @param chapter the chapter.
      * @return the pages for the chapter.
      */
-    fun fetchPageList(chapter: SChapter): Observable<List<Page>>
+    suspend fun getPageList(chapter: SChapter): List<Page>
 
-    @Deprecated(
-        "Use the non-RxJava API instead",
-        ReplaceWith("getMangaDetails"),
-    )
+    @Deprecated("Use the combined suspend API instead", ReplaceWith("getMangaUpdate"))
     fun fetchMangaDetails(manga: SManga): Observable<SManga> = throw Exception("Stub!")
 
-    @Deprecated(
-        "Use the non-RxJava API instead",
-        ReplaceWith("getChapterList"),
-    )
+    @Deprecated("Use the combined suspend API instead", ReplaceWith("getMangaUpdate"))
     fun fetchChapterList(manga: SManga): Observable<List<SChapter>> = throw Exception("Stub!")
+
+    @Deprecated("Use the suspend API instead", ReplaceWith("getPageList"))
+    fun fetchPageList(chapter: SChapter): Observable<List<Page>> = throw Exception("Stub!")
 }